third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/metrics-box-sizing-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that content-box and border-box content area dimensions are handled property by the Metrics pane.


Running: testBorderBoxInit1

Running: testInitialBorderBoxMetrics
=== Initial border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 35 x 35

Running: testModifiedBorderBoxMetrics
=== Modified border-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 35

Running: testContentBoxInit1

Running: testInitialContentBoxMetrics
=== Initial content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 55 x 55

Running: testModifiedContentBoxMetrics
=== Modified content-box ===
margin: 1 1 1 1
border: 3 3 3 3
padding: 7 7 7 7
content: 60 x 55
